LayoutTests/accessibility/mac/text-marker-length-expected.txt

text block
This tests the text marker implemention of DRT is working by returning a marker range and by returning its correct length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.textMarkerRangeLength(textMarkerRange) is 10
PASS successfullyParsed is true

TEST COMPLETE
